test: fix import path (#25)
